+++ src/app/onboarding/_component/SignupHeader.tsx
