+++ src/app/onboarding/_component/funnel/FinishFunnel.tsx
