+++ src/app/onboarding/_component/funnel/SelectCategory.tsx
