+++ src/app/onboarding/_component/funnel/SelectResidence.tsx
